refactor: Consolidate transition duration
